Integrate AI waste classification and enhance UI
Fixed the camera based classifier with some "tips" on reuse and recycling of the object. Gives points as well, need to come up with a server that can store and track points for each user and also remove unnecessary elements like quick classify, change the broader categories and give the color of the bin to throw the waste into.